+++ Server/API/Controllers/PreferencesController.cs
@@ -1,60 +1,73 @@
-﻿using API.Filters;
+﻿using API.Domain;
+using API.Filters;
 using API.Interfaces.IServices;
 using Authorization;
 using Authorization.Resources;
+using IODomain.Extensions;
 using IODomain.Input;
-using IODomain.Output;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace API.Controllers {
+    [ValidateModel]
     [ServicesExceptionFilter]
     [Route("api/Users/{userId}/[controller]")]
     public class PreferencesController : Controller {
         private readonly IPreferencesService _preferencesService;
-        private readonly IUserService _userService;
-        private readonly IUsersBoardsService _usersBoardsService;
         private readonly IAuthorizationService _authorizationService;
 
-        public PreferencesController(IPreferencesService preferencesService, IUserService userService, IUsersBoardsService usersBoardsService, IAuthorizationService authorizationService) {
+        public PreferencesController(IPreferencesService preferencesService, IAuthorizationService authorizationService) {
             _preferencesService = preferencesService;
-            _userService = userService;
-            _usersBoardsService = usersBoardsService;
             _authorizationService = authorizationService;
         }
 
         [HttpGet(Name = "GetPreferences")]
         public async Task<IActionResult> GetById(string userId) {
             if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(userId), Policies.UserIsOwnPolicy))
-                return new ChallengeResult();
+                return Challenge();
 
-            OutPreferences outPreferences = await _preferencesService.GetAsync(userId);
+            Preferences preferences = await _preferencesService.GetAsync(userId);
 
-            return new OkObjectResult(outPreferences);
+            if(preferences == null) {
+                //TODO Ver o que fazer
+                //Ideia fornecer default preferences
+            }
+
+            return Ok(preferences.Out());
         }
 
         [HttpPut]
         public async Task<IActionResult> CreateOrUpdate(string userId, [FromBody] InPreferences inPreferences) {
             if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(userId), Policies.UserIsOwnPolicy))
-                return new ChallengeResult();
+                return Challenge();
 
-            OutPreferences resourceCreated = await _preferencesService.CreateOrUpdateAsync(userId, inPreferences);
+            if(inPreferences.UserId != userId) {
+                return BadRequest();
+            }
 
-            if(resourceCreated != null)
-                return CreatedAtRoute("GetPreferences", new { userId = userId }, resourceCreated);
-            
-            return new NoContentResult();
+            Preferences preferences = new Preferences { UserId = userId }.In(inPreferences);
+
+            bool created = await _preferencesService.CreateOrUpdateAsync(preferences);
+
+            if(created)
+                return CreatedAtRoute("GetPreferences", new { userId = userId }, preferences.Out());
+
+            return NoContent();
         }
 
         [HttpDelete]
         public async Task<IActionResult> Delete(string userId) {
             if(!await _authorizationService.AuthorizeAsync(User, new UserRequest(userId), Policies.UserIsOwnPolicy))
-                return new ChallengeResult();
+                return Challenge();
+
+            if(!await _preferencesService.ExistsAsync(userId)) {
+                return NotFound();
+            }
 
             await _preferencesService.DeleteAsync(userId);
-            return new NoContentResult();
+
+            return NoContent();
         }
     }
 }
